@@ -15,11 +15,12 @@ class ItemsCommandsDecoupledCtrl(Controller):
     path = "/items_decoupled"
     tags = ["Items Commands Decoupled"]
 
-
     @post(path="", dto=NewItemDTO)
     @inject
     async def post_item(
-        self, data: Item, item_command_publisher: FromDishka[AbstractItemCommandPublisher]
+        self,
+        data: Item,
+        item_command_publisher: FromDishka[AbstractItemCommandPublisher],
     ) -> None:
         await item_command_publisher.create_item(data)
         return
@@ -27,7 +28,9 @@ class ItemsCommandsDecoupledCtrl(Controller):
     @patch(path="", dto=UpdateItemDTO)
     @inject
     async def patch_item(
-        self, data: Item, item_command_publisher: FromDishka[AbstractItemCommandPublisher]
+        self,
+        data: Item,
+        item_command_publisher: FromDishka[AbstractItemCommandPublisher],
     ) -> None:
         await item_command_publisher.update_item(data)
         return
@@ -35,7 +38,9 @@ class ItemsCommandsDecoupledCtrl(Controller):
     @delete(path="/{item_id:int}")
     @inject
     async def delete_item(
-        self, item_id: int, item_command_publisher: FromDishka[AbstractItemCommandPublisher]
+        self,
+        item_id: int,
+        item_command_publisher: FromDishka[AbstractItemCommandPublisher],
     ) -> None:
         await item_command_publisher.delete_item(item_id)
         return
