@@ -32,7 +32,9 @@ async def update_item(item: Item, unit_of_work: FromDishka[AbstractUnitOfWork]) 
 
 
 @router.subscriber("delete_command")
-async def delete_item(item_id: int, unit_of_work: FromDishka[AbstractUnitOfWork]) -> None:
+async def delete_item(
+    item_id: int, unit_of_work: FromDishka[AbstractUnitOfWork]
+) -> None:
     with unit_of_work:
         success: bool = unit_of_work.items.delete(item_id)
         if not success:
