@@ -1,7 +1,7 @@
 import logging
 
 from app.litestar_app_factory import create_app
-    
+
 logging.basicConfig(
     level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s"
 )
