@@ -1,8 +1,9 @@
 from typing import Optional, TypeVar, Generic, List
 from abc import ABC, abstractmethod
 
-T = TypeVar('T')
-ID = TypeVar('ID')
+T = TypeVar("T")
+ID = TypeVar("ID")
+
 
 class AbstractRepository(ABC, Generic[T, ID]):
     @abstractmethod
