@@ -2,7 +2,7 @@ import os
 import sys
 import pytest
 
-if os.name == 'nt':
+if os.name == "nt":
     import msvcrt
 else:
     import fcntl
@@ -14,7 +14,7 @@ LOCKFILE_PATH = os.path.join(os.path.dirname(__file__), ".pytest_lock")
 def lock_test():
     """Cross-platform, process-safe lock fixture for serializing tests."""
     with open(LOCKFILE_PATH, "w") as lockfile:
-        if os.name == 'nt':
+        if os.name == "nt":
             msvcrt.locking(lockfile.fileno(), msvcrt.LK_LOCK, 1)
         else:
             fcntl.flock(lockfile, fcntl.LOCK_EX)
@@ -22,7 +22,7 @@ def lock_test():
         try:
             yield
         finally:
-            if os.name == 'nt':
+            if os.name == "nt":
                 msvcrt.locking(lockfile.fileno(), msvcrt.LK_UNLCK, 1)
             else:
                 fcntl.flock(lockfile, fcntl.LOCK_UN)
