@@ -5,15 +5,18 @@ from litestar.testing import TestClient
 from app.utils.lock_test import lock_test
 from app.items.db.fake_item_repository import FakeItemRepository
 
+
 @pytest.mark.separate
 @pytest.mark.integration
 class TestItemDecoupledCtrlIntegration:
     @pytest.fixture(autouse=True)
     def lock(self, lock_test):
         pass
-    
+
     def test_post_item(
-        self, fixture_integration_test_client_with_auth: TestClient, fixture_new_item: dict
+        self,
+        fixture_integration_test_client_with_auth: TestClient,
+        fixture_new_item: dict,
     ):
         client = fixture_integration_test_client_with_auth
 
@@ -28,18 +31,23 @@ class TestItemDecoupledCtrlIntegration:
         assert response_json[0]["Id"] is not None
 
     def test_delete_item(
-        self, fixture_integration_test_client_with_auth: TestClient, fixture_new_item: dict
+        self,
+        fixture_integration_test_client_with_auth: TestClient,
+        fixture_new_item: dict,
     ):
         response = fixture_integration_test_client_with_auth.post(
-            "/items_decoupled", json=fixture_new_item  # Pass the dictionary directly
+            "/items_decoupled",
+            json=fixture_new_item,  # Pass the dictionary directly
         )
         assert response.is_success
-        
+
         response = fixture_integration_test_client_with_auth.get("/items")
         assert response.is_success
         item_id = response.json()[0]["Id"]
 
-        response = fixture_integration_test_client_with_auth.delete(f"/items_decoupled/{item_id}")
+        response = fixture_integration_test_client_with_auth.delete(
+            f"/items_decoupled/{item_id}"
+        )
         assert response.is_success
 
         response = fixture_integration_test_client_with_auth.get("/items")
@@ -53,7 +61,8 @@ class TestItemDecoupledCtrlIntegration:
         fixture_update_item: dict,
     ):
         response = fixture_integration_test_client_with_auth.post(
-            "/items_decoupled", json=fixture_new_item  # Pass the dictionary directly
+            "/items_decoupled",
+            json=fixture_new_item,  # Pass the dictionary directly
         )
         assert response.is_success
 
@@ -64,7 +73,8 @@ class TestItemDecoupledCtrlIntegration:
         update_item = fixture_update_item
         update_item["Id"] = item_id  # Modify the update item to include the ID
         response = fixture_integration_test_client_with_auth.patch(
-            "/items_decoupled", json=update_item  # Send the updated item as JSON
+            "/items_decoupled",
+            json=update_item,  # Send the updated item as JSON
         )
         assert response.is_success
 
